backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
numpy
scipy
PyMySQL
gunicorn
redis
prometheus-client
